moved services to separate namespace
